redo validators
