+++ packages/components/src/validators/BaseValidator.ts
@@ -1,10 +1,9 @@
 import { TranslatorType } from '@ids-context/Translator';
 
-// eslint-disable-next-line @typescript-eslint/no-unnecessary-type-parameters
 export default abstract class BaseValidator<T> {
-    protected _translator!: TranslatorType;
+    protected _translator: TranslatorType;
 
-    setTranslator(translator: TranslatorType) {
+    constructor(translator: TranslatorType) {
         this._translator = translator;
     }
